[Refactor] 스크린샷을 위한 더미로직 미묘한 수정, static Bool 빼기 전 로직
